CCSSSCON-353: Frame out server gateway implementation for DXF extract boundaries service

+++ src/service/TRex/src/netstandard/VSS.TRex.Files/DXF/PolyLineBoundaries.cs
@@ -8,10 +8,10 @@ namespace VSS.TRex.Files.DXF
     public DxfUnitsType Units;
     public List<PolyLineBoundary> Boundaries;
 
-    public PolyLineBoundaries(DxfUnitsType units, int maxBoundariesToProcess)
+    public PolyLineBoundaries(DxfUnitsType units, uint maxBoundariesToProcess)
     {
       Units = units;
-      Boundaries = new List<PolyLineBoundary>(maxBoundariesToProcess);
+      Boundaries = new List<PolyLineBoundary>((int)maxBoundariesToProcess);
     }
   }
 }
